Refactor screen
