+++ README.md
@@ -3,3 +3,9 @@
 This mod adds a new heath system to Minecraft.
 Players health is splitted in different body parts, when a part receives damage only the armor corresponding to that part is taken in consideration.
 If some parts lose all the health debuffs will be applied to the player, as slowness, mining fatigue and weakness.
+
+## TODO
+
+- Ray cast arrows to apply the damage to a specific part.
+- HUD config
+- Beter GUI
