+++ src/main/java/xyz/srgnis/bodyhealthsystem/registry/CustomScreen.java
@@ -1,7 +1,7 @@
 package xyz.srgnis.bodyhealthsystem.registry;
 
 import net.minecraft.client.gui.screen.ingame.HandledScreens;
-import xyz.srgnis.bodyhealthsystem.client.HealScreen;
+import xyz.srgnis.bodyhealthsystem.client.screen.HealScreen;
 
 public class CustomScreen {
     public static void registerScreens() {
