+++ src/main/java/xyz/srgnis/bodyhealthsystem/registry/CustomScreenHandler.java
@@ -4,7 +4,7 @@ import net.minecraft.screen.ScreenHandlerType;
 import net.minecraft.util.Identifier;
 import net.minecraft.util.registry.Registry;
 import xyz.srgnis.bodyhealthsystem.BHSMain;
-import xyz.srgnis.bodyhealthsystem.client.HealScreenHandler;
+import xyz.srgnis.bodyhealthsystem.client.screen.HealScreenHandler;
 
 public class CustomScreenHandler {
     public static ScreenHandlerType<? extends HealScreenHandler> BAG_SCREEN_HANDLER;
